fix(frontend): address peqy review on test drawer rendering
Use unknown[] for turn.context, preserve plain-text newlines with
pre-wrap, and only render MarkdownContent when markdown syntax is detected.
